+++ R/ffs_demo.R
@@ -44,33 +44,33 @@
 #' ##
 #' ## without births and deaths over period
 #' ##
-#' # data as in papers
+#' # data as in demographic research and science paper papers
+#' 
 #' s1 <- matrix(data = c(1000, 100, 10, 0, 55, 555, 50, 5, 80, 40, 800, 40, 20, 25, 20, 200),
 #'              nrow = 4, ncol = 4, byrow = TRUE)
 #' s2 <- matrix(data = c(950, 100, 60, 0, 80, 505, 75, 5, 90, 30, 800, 40, 40, 45, 0, 180),
 #'              nrow = 4, ncol = 4, byrow = TRUE)
 #' b <- d <- rep(0, 4)
 #' r <- LETTERS[1:4]
-#' dimnames(s1) <- dimnames(s2) <- list(pob = r, por = r)
+#' dimnames(s1) <- dimnames(s2) <- list(birth =  r, dest = r)
 #' names(b) <- names(d) <- r
-#' 
-#' s1
-#' s2
+#' addmargins(s1)
+#' addmargins(s2)
 #' b
 #' d
 #' 
 #' # demographic research and science paper example
 #' e0 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)
-#' e0$od_flow
+#' e0$od
 #' 
 #' # international migration review paper example
 #' s1[,] <- c(100, 20, 10, 20, 10, 55, 40, 25, 10, 25, 140, 20, 0, 10, 65, 200)
 #' s2[,] <- c(70, 25, 10, 40, 30, 60, 55, 45, 10, 10, 140, 0, 10, 15, 50, 180)
-#' s1
-#' s2
+#' addmargins(s1)
+#' addmargins(s2)
 #' 
 #' e1 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)
-#' e1$od_flow
+#' e1$od
 #' 
 #' # international migration review supp. material example
 #' # distance matrix
@@ -79,7 +79,7 @@
 #' dimnames(dd) <- list(orig = r, dest = r)
 #' dd
 #' e2 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d, m = dd)
-#' e2$od_flow
+#' e2$od
 #' 
 #' ##
 #' ## with births and deaths over period
@@ -90,14 +90,15 @@
 #' b[] <- c(80, 20, 40, 60)
 #' d[] <- c(70, 30, 50, 10)
 #' e3 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d, match_pob_tot_method = "open-dr")
+#' e3$od
 #' # makes more sense to use this method
 #' e4 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d, match_pob_tot_method = "open")
-#' e4$od_flow
+#' e4$od
 #' 
 #' # science paper  supp. material example
 #' b[] <- c(80, 20, 60, 60)
 #' e5 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)
-#' e5$od_flow
+#' e5$od
 #' 
 #' # international migration review supp. material example (with births and deaths)
 #' s1[,] <- c(100, 20, 10, 20, 10, 55, 40, 25, 10, 25, 140, 20, 0, 10, 65, 200)
@@ -106,9 +107,20 @@
 #' d[] <- c(30, 10, 40, 10)
 #' e6 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)
 #' e6$od_flow
+#' 
+#' # scientific data 2019 paper
+#' s1[] <- c(100, 80, 30, 60, 10, 180, 10, 70, 10, 10, 140, 10, 0, 90, 40, 160)
+#' s2[] <- c(95, 75, 55, 35, 5, 225, 0, 25, 15, 5, 115, 25, 5, 55, 50, 215)
+#' b[] <- c(0, 0, 0, 0)
+#' d[] <- c(0, 0, 0, 0)
+#' e7 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)
+#' e7$od
+# source("./R/ipf_seed.R")
+# source("./R/birth_mat.R"); source("./R/death_mat.R")
+# source("./R/match_pob_tot.R"); source("./R/rescale_nb.R")
 # m1 = s1; m2 = s2; b_por = b; d_por = d; m = NULL
 # m1 = s1; m2 = s2; b_por = births; d_por = deaths; m = NULL
-# stayer_assumption = TRUE; match_pob_tot_method = "rescale"; birth_non_negative = TRUE; death_method = "proportion"; match_global = "after-demo-adjust"; verbose = FALSE
+# stayer_assumption = TRUE; match_pob_tot_method = "rescale"; birth_non_negative = TRUE; birth_method = "native"; death_method = "proportion"; match_global = "after-demo-adjust"; verbose = FALSE
 # birth_method = "proportion"
 # match_pob_tot_method = "open";
 ffs_demo <- function(m1 = NULL,
@@ -125,8 +137,8 @@ ffs_demo <- function(m1 = NULL,
                      verbose = FALSE,
                      ...) {
   # make sure dimensions match
-  R <- unique(c(dim(m1), dim(m2), length(b_por), length(d_por)))
-  if (length(R) != 1)
+  n <- unique(c(dim(m1), dim(m2), length(b_por), length(d_por)))
+  if (length(n) != 1)
     stop("m1 and m2 matrices must be square and with the same dimension as length of b_por and d_por.")
   
   # make sure data in same order
@@ -134,42 +146,38 @@ ffs_demo <- function(m1 = NULL,
     stop("m1 and m2 must be named matrices, b_por and d_por must be named vectors, all names must be common")
   
   # get dimension names and everything in same order
-  dn <- union(names(b_por), names(d_por))
-  dn <- union(dn, dimnames(m1)[[1]])
-  dn <- union(dn, dimnames(m1)[[2]])
-  dn <- union(dn, dimnames(m2)[[1]])
-  dn <- union(dn, dimnames(m2)[[2]])
-  m1_a <- m1[dn, dn]
-  m2_a <- m2[dn, dn]
-  b <- b_por[dn]
-  d <- d_por[dn]
+  r <- union(names(b_por), names(d_por))
+  r <- union(r, dimnames(m1)[[1]])
+  r <- union(r, dimnames(m1)[[2]])
+  r <- union(r, dimnames(m2)[[1]])
+  r <- union(r, dimnames(m2)[[2]])
+  m1_a <- m1[r, r]
+  m2_a <- m2[r, r]
+  b <- b_por[r]
+  d <- d_por[r]
   
   # set up m and y to store results
   if(is.null(m) | length(dim(m) == 2)){
-    m <- ipf_seed(m = m, R = R, n_dim = 3, dn = dn)
-    # m <- migest:::ipf_seed(m = m, R = R, n_dim = 3, dn = dn)
+    m <- ipf_seed(m = m, R = n, n_dim = 3, dn = r)
   }
   y <- array(0, dim(m) + c(2, 2, 0))
-  dimnames(y) <- list(orig = c(dn, "birth", "outside"),
-                      dest = c(dn, "death", "outside"),
-                      pob = dn)
+  dimnames(y) <- list(orig = c(r, "birth", "outside"),
+                      dest = c(r, "death", "outside"),
+                      pob = r)
   
   if(match_global == "before-demo-adjust"  & match_pob_tot_method %in% c("rescale", "rescale-adjust-zero-fb")){
     if(verbose)
       message("Rescale native born cells for global zero net migration...")
-    x <- rescale_nb(m1 = m1_a, m2 = m2_a, b = b, d = d)
-    m1_a <- x$m1_adj
-    m2_a <- x$m2_adj
+    x1 <- rescale_nb(m1 = m1_a, m2 = m2_a, b = b, d = d)
+    m1_a <- x1$m1_adj
+    m2_a <- x1$m2_adj
   }
   
   # adjust for births and deaths
   if(verbose)
     message("Adjust stock tables for changes in births and deaths...")
-  
   b_mat <- birth_mat(b_por = b, m2 = m2_a, method = birth_method, non_negative = birth_non_negative)
   d_mat <- death_mat(d_por = d, m1 = m1_a, method = death_method, m2 = m2_a, b_por = b)
-  # b_mat <- migest:::birth_mat(b_por = b, m2 = m2_a, method = birth_method, non_negative = birth_non_negative)
-  # d_mat <- migest:::death_mat(d_por = d, m1 = m1_a, method = death_method, m2 = m2_a, b_por = b)
   m1_b <- m1_a - d_mat
   m2_b <- m2_a - b_mat
   
@@ -189,37 +197,65 @@ ffs_demo <- function(m1 = NULL,
   # adjust for pob rows to match
   if(verbose)
     message("Rescale stock tables for equal place of birth totals...")
-  x <- match_pob_tot(m1 = m1_c, m2 = m2_c, method = match_pob_tot_method, verbose = verbose)
-  # x <- migest:::match_pob_tot(m1 = m1_c, m2 = m2_c, method = match_pob_tot_method, verbose = verbose)
-  m1_d <- x$m1_adj
-  m2_d <- x$m2_adj
-  # round(rowSums(m1_d)); round(rowSums(m2_d));
-  
+  x2 <- match_pob_tot(m1 = m1_c, m2 = m2_c, method = match_pob_tot_method, verbose = verbose)
+  m1_d <- x2$m1_adj
+  m2_d <- x2$m2_adj
+
   # ipf
   if(verbose)
     message("Estimate flows to match changes in adjusted stocks")
-  if(stayer_assumption)
-    fl <- ipf3_qi(row_tot = t(m1_d), col_tot = m2_d, m = m, verbose = verbose, ...)$mu
-  if(!stayer_assumption)
-    fl <- ipf3(row_tot = t(m1_d), col_tot = m2_d, m = m, verbose = verbose, ...)$mu
-    # fl <- mipfp::Ipfp(seed = m, tol = 1e-03, iter = 1e05, 
-    #                   # print = TRUE, 
-    #                   target.list = list(c(1, 3), c(2,3)),
-    #                   target.data = list(t(m1_d), m2_d))$x.hat
+  if(stayer_assumption){
+    d0 <- expand.grid(a = 1:n, b = 1:n)
+    diag_count <- diag_zero <- array(0, c(n, n, n))
+    diag_count <- with(
+      data = d0, 
+      expr = replace(x = diag_count,
+                     list = cbind(a, a, b), 
+                     values = apply(X = cbind(c(t(m1_d)), c(t(m2_d))), 
+                                    MARGIN = 1, 
+                                    FUN = min)))
+    dimnames(diag_count) <- dimnames(diag_zero) <- dimnames(m)
+    
+    diag_zero <- with(data = d0, expr = replace(x = diag_zero, 
+                                                list = cbind(a, a, b), 
+                                                values = 0))
+    if(sum(m) != n*n*3)
+      diag_zero <- m
+    a0 <- mipfp::Ipfp(
+      seed = diag_zero, tol = 1e-03, iter = 1e05,
+      # print = TRUE,
+      target.list = list(c(1, 3), c(2,3)),
+      target.data = list(t(m1_d) - apply(X = diag_count, MARGIN = c(1, 3), FUN = sum),
+                         t(m2_d) - apply(X = diag_count, MARGIN = c(1, 3), FUN = sum)))
+    f0 <- a0$x.hat + diag_count
+  }
 
+  if(!stayer_assumption){
+    # fl <- ipf3(row_tot = t(m1_d), col_tot = m2_d, m = m, verbose = verbose, ...)$mu
+    f0 <- mipfp::Ipfp(seed = m, tol = 1e-03, iter = 1e05,
+                      # print = TRUE,
+                      target.list = list(c(1, 3), c(2,3)),
+                      target.data = list(t(m1_d), m2_d))$x.hat
+    
+  }
+    
   # fill in y
-  y[1:R, 1:R, ] <- fl
-  y[R + 1, 1:R, ] <- b_mat
-  y[R + 2, 1:R, ] <- t(x$out_mat)
-  y[1:R, R + 1, ] <- t(d_mat)
-  y[1:R, R + 2, ] <- t(x$in_mat)
+  y[1:n, 1:n, ] <- f0
+  y[n + 1, 1:n, ] <- b_mat
+  y[n + 2, 1:n, ] <- t(x2$out_mat)
+  y[1:n, n + 1, ] <- t(d_mat)
+  y[1:n, n + 2, ] <- t(x2$in_mat)
   
-  return(list(flow = fl, 
-              y = y, 
-              stayer_assumption = stayer_assumption,
-              match_pob_tot_method = match_pob_tot_method,
-              birth_non_negative = birth_non_negative,
-              death_method = death_method, 
-              od_flow = stats::addmargins(sum_od(fl))
-  ))
+  return(
+    list(flow = f0, 
+         y = y, 
+         stayer_assumption = stayer_assumption,
+         match_pob_tot_method = match_pob_tot_method,
+         birth_non_negative = birth_non_negative,
+         death_method = death_method, 
+         od_flow = f0 %>%
+           sum_od() %>%
+           stats::addmargins()
+    )
+  )
 }