+++ NEWS.md
@@ -3,6 +3,7 @@
 * Dropped `sum_turnover()` and `sum_bilateral()`
 * Fixed axis upper limits in `mig_chord()`
 * Added `region_ac2022` and `region_wb` column to `dict_ims`
+* Replaced `ipf` functions in `ffs_demo` and sub-functions with `mipfp::Ipfp`
 
 # migest 2.0.2